refactor(scripts): extract shared npm-registry helper from publish-sdk
Pull the npm-view classification (version resolves vs 404 vs real error),
the spawnSync command runner, and the CommandRunner/CommandResult types into
scripts/lib/npm-registry.ts so release tooling shares one source of truth.
publish-sdk.ts imports them and re-exports CommandRunner; behavior unchanged.
